+++ src/linalg/eig_power_d.cc
@@ -32,7 +32,7 @@ double eig_power_right(const RTensor &O, RTensor *vector, size_t iter,
                        double tol) {
   assert(O.rows() == O.columns());
   return eig_power([&O](const RTensor &x) -> RTensor { return mmult(O, x); },
-                   O.columns(), vector, iter, tol);
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 /**Left eigenvalue and eigenvector with the largest absolute
@@ -47,7 +47,7 @@ double eig_power_left(const RTensor &O, RTensor *vector, size_t iter,
   assert(O.rows() == O.columns());
   auto OT = transpose(O);
   return eig_power([&OT](const RTensor &x) -> RTensor { return mmult(OT, x); },
-                   O.columns(), vector, iter, tol);
+                   static_cast<size_t>(O.columns()), vector, iter, tol);
 }
 
 }  // namespace linalg
